数据前后处理由子类单独实现
