+++ lazyllm/module/llms/onlinemodule/supplier/doubao.py
@@ -48,6 +48,15 @@ class DoubaoEmbedding(OnlineEmbeddingModuleBase):
                 for i in range(len(json_data)):
                     json_data[i].update(kwargs)
             return json_data
+        
+    def _parse_response(self, response: Dict, input: Union[List, str]) -> Union[List[List[float]], List[float]]:
+        data = response.get('data', [])
+        if not data:
+            raise Exception('no data received')
+        if isinstance(input, str):
+            return data[0].get('embedding', [])
+        else:
+            return [res.get('embedding', []) for res in data]        
 
 class DoubaoMultimodalEmbedding(OnlineEmbeddingModuleBase):
     def __init__(self,
